Reduce thread pool depletion issues on modern .NET
Motivation
----------
.NET 6 and later includes special logic to recognize when a thread pool
thread is blocked waiting on a Task to complete, which is generally
also waiting on a thread pool thread. When safe to do so, we should
block on the task directly rather than using
ExclusiveSynchronizationContext so that .NET can recognize the case and
grow the thread pool more aggressively. This also reduces other overhead
as well, improving overall performance.

Modifications
-------------
When RunSync is called with no or the default SynchronizationContext and
no custom TaskScheduler, just block directly on the returned awaiter
rather than constructing an ExclusiveSynchronizationContext.