Provide visibility to running synchronously (#16)
Motivation
----------
There are cases in complex systems where it is helpful for code to know that it is running within a call to AsyncHelper.RunSync. Furthermore, retrieving the previously replaced synchronization context can also be helpful. For example, a call to Form.ShowDialog in a WinForms app may want to reinstall the WindowsFormsSynchronizationContext since a new message pump will be running to process continuations.

Modifications
-------------
- Add IsRunningSynchronously property
- Add GetReplacedSynchronizationContext method
- Track the fact we're in the call to RunSync and processing work on its calling thread using a thread static